third_party/WebKit/LayoutTests/media/media-can-play-webm-expected.txt

Test HTMLMediaElement canPlayType() method for webm media containers.

These tests may be expected to fail if the WebKit port does not support the format.

EXPECTED (video.canPlayType('audio/webm') == 'maybe') OK
EXPECTED (video.canPlayType('video/webm') == 'maybe') OK
EXPECTED (video.canPlayType('audio/webm; codecs=vorbis') == 'probably') OK
EXPECTED (video.canPlayType('video/webm; codecs=vp8,vorbis') == 'probably') OK
EXPECTED (video.canPlayType('video/webm; codecs=vp8') == 'probably') OK
EXPECTED (video.canPlayType('video/webm; codecs=vp8.0') == 'probably') OK
EXPECTED (video.canPlayType('video/webm; codecs=vp8.0,vorbis') == 'probably') OK
EXPECTED (video.canPlayType('video/webm; codecs=vp8.1') == '') OK
EXPECTED (video.canPlayType('video/webm; codecs=vp9') == 'probably') OK
EXPECTED (video.canPlayType('video/webm; codecs=vp9,vorbis') == 'probably') OK
EXPECTED (video.canPlayType('video/webm; codecs=vp9.0') == 'probably') OK
EXPECTED (video.canPlayType('video/webm; codecs=vp9.0,vorbis') == 'probably') OK
EXPECTED (video.canPlayType('video/webm; codecs=vp9.1') == '') OK
EXPECTED (video.canPlayType('audio/webm; codecs=opus') == 'probably') OK
EXPECTED (video.canPlayType('video/webm; codecs=vp9,opus') == 'probably') OK
EXPECTED (video.canPlayType('video/webm; codecs=vp9.0,opus') == 'probably') OK
EXPECTED (video.canPlayType('video/webm; codecs=vp8,opus') == 'probably') OK
EXPECTED (video.canPlayType('video/webm; codecs=vp8.0,opus') == 'probably') OK
END OF TEST
